backend/requirements.txt
# INEC Underage Eradicator - Requirements
# Python dependencies for the voter registration system

# Django and web framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary>=2.9.9

# AI and ML libraries
opencv-python>=4.8.1.78
Pillow>=10.1.0
deepface>=0.0.79
# insightface>=0.7.3
tensorflow>=2.15.0
scikit-learn>=1.3.2
# pandas>=2.1.4
numpy>=1.26.2

# OCR libraries
pytesseract>=0.3.10
easyocr>=1.7.1
# paddlepaddle>=2.5.1
# paddleocr>=2.7.0

# PDF generation
reportlab>=4.0.7
weasyprint>=61.2

# Security and encryption
cryptography>=41.0.7
bcrypt>=4.1.2

# Data generation and testing
faker>=20.1.0

# Image processing
scikit-image>=0.22.0

# QR codes
qrcode[pil]>=7.4.2

# Caching and performance
redis>=5.0.1
django-redis>=5.4.0

# Logging and monitoring
structlog>=23.2.0

# Development and testing
pytest>=7.4.3
pytest-django>=4.7.0
coverage>=7.3.2

# Web server
gunicorn>=21.2.0
whitenoise>=6.6.0

# API documentation
drf-spectacular>=0.26.5

# Email (if needed)
# django-anymail==10.2

# Task queue (if needed)
# celery==5.3.4

# Monitoring (if needed)
# sentry-sdk==1.38.0